Implement M4->M0 message queue
Remove the current phase code.

Implement a queue so that M0 can post messages into the queue and M0 can
pull the messages out. The messages encode a buffer offset and length.
This allows M4 to tell M0 to transmit X bytes from offset Y over the USB
bus.

The advantage of this is that the M0 code can be common for all
byte/word packing carried out by M4. Also up to 3 commands can be queued
allowing M0 to catch up if it is slow to transmit over the USB.